rag_pipeline/requirements_rag.txt
# RAG 向量检索依赖
# pip install -r rag_pipeline/requirements_rag.txt -i https://mirrors.tuna.tsinghua.edu.cn/pypi/web/simple

# 向量库
qdrant-client>=1.9

# 嵌入模型（本地推理，GPU 可选）
sentence-transformers>=3.0
# GPU 支持：pip install torch --index-url https://download.pytorch.org/whl/cu121

# Reranker（可选，精度更高）
FlagEmbedding>=1.2

# BM25 稀疏检索
rank_bm25>=0.2
jieba>=0.42

# 已有依赖（保留）
python-dotenv>=1.0
openai>=1.0
